# perf(gatsby-source-contentful): dont re-create nodes