… packages/gatsby-source-contentful/src/__tests__/__snapshots__/normalize.js.snap