@wardpeet 2020-12-15
This makes code harder to reason with. The real issue is inside gatsby-node where we merge existing nodes together with new nodes. That part needs to refactor so I'm inclined to close that PR for that reason