@wardpeet 2020-12-16
I was missing a lot of context on the why here. I'm fine with this patch as it does improve performance. If we could run some tests to make sure references and all are still being updated 👍