@zcei 2020-12-17
Heja there 👋 

got hinted to this PR by @axe312ger, I'm one of Contentful's backend engineers and want to increase your confidence into the solution. 🙂 

For a single given Contentful entity, a write towards it will update its `.sys.updatedAt` property. This is even the case for no-change writes (i.e. a PUT to an entity with the same payload as its current shape).

That means using the `.sys.updatedAt` property for cache invalidation even gives users the ability to easily invalidate single nodes without changing content (why ever this might be useful 🤷‍♂️ ).

Phrased differently: 
* **same updatedAt** the content of two versions of one entity _is_ the same.
* **different updatedAt** the content of two versions of one entity _might_ be the same 

Hope this helps!